Fix unittests

+++ internal/schema/migrations.go
@@ -18,7 +18,7 @@ import (
 
 // migrationList returns a list of migrations to be executed. If the id of the
 // migration already exists in the migrations table it will be skipped.
-func migrationList(db *sqlx.DB, log *log.Logger) []*sqlxmigrate.Migration {
+func migrationList(db *sqlx.DB, log *log.Logger, isUnittest bool) []*sqlxmigrate.Migration {
 	return []*sqlxmigrate.Migration{
 		// Create table users.
 		{
@@ -249,18 +249,22 @@ func migrationList(db *sqlx.DB, log *log.Logger) []*sqlxmigrate.Migration {
 					return errors.WithMessagef(err, "Failed to prepare sql query '%s'", q)
 				}
 
-				resChan := make(chan interface{})
-				go geonames.LoadGeonames(context.Background(), resChan)
+				if isUnittest {
 
-				for r := range resChan {
-					switch v := r.(type) {
-					case geonames.Geoname:
-						_, err = stmt.Exec(v.CountryCode, v.PostalCode, v.PlaceName, v.StateName, v.StateCode, v.CountyName, v.CountyCode, v.CommunityName, v.CommunityCode, v.Latitude, v.Longitude, v.Accuracy)
-						if err != nil {
-							return errors.WithStack(err)
+				} else {
+					resChan := make(chan interface{})
+					go geonames.LoadGeonames(context.Background(), resChan)
+
+					for r := range resChan {
+						switch v := r.(type) {
+						case geonames.Geoname:
+							_, err = stmt.Exec(v.CountryCode, v.PostalCode, v.PlaceName, v.StateName, v.StateCode, v.CountyName, v.CountyCode, v.CommunityName, v.CommunityCode, v.Latitude, v.Longitude, v.Accuracy)
+							if err != nil {
+								return errors.WithStack(err)
+							}
+						case error:
+							return v
 						}
-					case error:
-						return v
 					}
 				}
 
@@ -287,9 +291,35 @@ func migrationList(db *sqlx.DB, log *log.Logger) []*sqlxmigrate.Migration {
 			ID: "20190731-02d",
 			Migrate: func(tx *sql.Tx) error {
 
-				prep := []string{
-					`DROP TABLE IF EXISTS countryinfo`,
-					`CREATE TABLE countryinfo (
+				schemas := []string{
+					// Countries...
+					`DROP TABLE IF EXISTS countries`,
+					`CREATE TABLE countries(
+						code           char(2) not null constraint countries_pkey primary key,
+						iso_alpha3           char(3),
+						name    character varying(50),
+						capital    character varying(50),
+						currency_code        char(3),
+						currency_name        CHAR(20),
+						phone                character varying(20),
+						postal_code_format        character varying(200),
+						postal_code_regex         character varying(200))`,
+				}
+
+				for _, q := range schemas {
+					_, err := db.Exec(q)
+					if err != nil {
+						return errors.WithMessagef(err, "Failed to execute sql query '%s'", q)
+					}
+				}
+
+				if isUnittest {
+					// `insert into countries(code, iso_alpha3, name, capital, currency_code, currency_name, phone, postal_code_format, postal_code_regex)
+
+				} else {
+					prep := []string{
+						`DROP TABLE IF EXISTS countryinfo`,
+						`CREATE TABLE countryinfo (
 						iso_alpha2           char(2),
 						iso_alpha3           char(3),
 						iso_numeric          integer,
@@ -309,38 +339,108 @@ func migrationList(db *sqlx.DB, log *log.Logger) []*sqlxmigrate.Migration {
 						geonameId            int,
 						neighbours           character varying(50),
 						equivalent_fips_code character varying(3))`,
-				}
+					}
 
-				for _, q := range prep {
-					_, err := db.Exec(q)
+					for _, q := range prep {
+						_, err := db.Exec(q)
+						if err != nil {
+							return errors.WithMessagef(err, "Failed to execute sql query '%s'", q)
+						}
+					}
+
+					u := "http://download.geonames.org/export/dump/countryInfo.txt"
+					resp, err := pester.Get(u)
 					if err != nil {
-						return errors.WithMessagef(err, "Failed to execute sql query '%s'", q)
+						return errors.WithMessagef(err, "Failed to read country info from '%s'", u)
 					}
-				}
+					defer resp.Body.Close()
 
-				u := "http://download.geonames.org/export/dump/countryInfo.txt"
-				resp, err := pester.Get(u)
-				if err != nil {
-					return errors.WithMessagef(err, "Failed to read country info from '%s'", u)
-				}
-				defer resp.Body.Close()
+					scanner := bufio.NewScanner(resp.Body)
+					var prevLine string
+					var stmt *sql.Stmt
+					for scanner.Scan() {
+						line := scanner.Text()
 
-				scanner := bufio.NewScanner(resp.Body)
-				var prevLine string
-				var stmt *sql.Stmt
-				for scanner.Scan() {
-					line := scanner.Text()
+						// Skip comments.
+						if strings.HasPrefix(line, "#") {
+							prevLine = line
+							continue
+						}
 
-					// Skip comments.
-					if strings.HasPrefix(line, "#") {
-						prevLine = line
-						continue
-					}
+						// Pull the last comment to load the fields.
+						if stmt == nil {
+							prevLine = strings.TrimPrefix(prevLine, "#")
+							r := csv.NewReader(strings.NewReader(prevLine))
+							r.Comma = '\t' // Use tab-delimited instead of comma <---- here!
+							r.FieldsPerRecord = -1
 
-					// Pull the last comment to load the fields.
-					if stmt == nil {
-						prevLine = strings.TrimPrefix(prevLine, "#")
-						r := csv.NewReader(strings.NewReader(prevLine))
+							lines, err := r.ReadAll()
+							if err != nil {
+								return errors.WithStack(err)
+							}
+							var columns []string
+
+							for _, fn := range lines[0] {
+								var cn string
+								switch fn {
+								case "ISO":
+									cn = "iso_alpha2"
+								case "ISO3":
+									cn = "iso_alpha3"
+								case "ISO-Numeric":
+									cn = "iso_numeric"
+								case "fips":
+									cn = "fips_code"
+								case "Country":
+									cn = "country"
+								case "Capital":
+									cn = "capital"
+								case "Area(in sq km)":
+									cn = "areainsqkm"
+								case "Population":
+									cn = "population"
+								case "Continent":
+									cn = "continent"
+								case "tld":
+									cn = "tld"
+								case "CurrencyCode":
+									cn = "currency_code"
+								case "CurrencyName":
+									cn = "currency_name"
+								case "Phone":
+									cn = "phone"
+								case "Postal Code Format":
+									cn = "postal_format"
+								case "Postal Code Regex":
+									cn = "postal_regex"
+								case "Languages":
+									cn = "languages"
+								case "geonameid":
+									cn = "geonameId"
+								case "neighbours":
+									cn = "neighbours"
+								case "EquivalentFipsCode":
+									cn = "equivalent_fips_code"
+								default:
+									return errors.Errorf("Failed to map column %s", fn)
+								}
+								columns = append(columns, cn)
+							}
+
+							placeholders := []string{}
+							for i := 0; i < len(columns); i++ {
+								placeholders = append(placeholders, "?")
+							}
+
+							q := "insert into countryinfo (" + strings.Join(columns, ",") + ") values(" + strings.Join(placeholders, ",") + ")"
+							q = db.Rebind(q)
+							stmt, err = db.Prepare(q)
+							if err != nil {
+								return errors.WithMessagef(err, "Failed to prepare sql query '%s'", q)
+							}
+						}
+
+						r := csv.NewReader(strings.NewReader(line))
 						r.Comma = '\t' // Use tab-delimited instead of comma <---- here!
 						r.FieldsPerRecord = -1
 
@@ -348,117 +448,36 @@ func migrationList(db *sqlx.DB, log *log.Logger) []*sqlxmigrate.Migration {
 						if err != nil {
 							return errors.WithStack(err)
 						}
-						var columns []string
 
-						for _, fn := range lines[0] {
-							var cn string
-							switch fn {
-							case "ISO":
-								cn = "iso_alpha2"
-							case "ISO3":
-								cn = "iso_alpha3"
-							case "ISO-Numeric":
-								cn = "iso_numeric"
-							case "fips":
-								cn = "fips_code"
-							case "Country":
-								cn = "country"
-							case "Capital":
-								cn = "capital"
-							case "Area(in sq km)":
-								cn = "areainsqkm"
-							case "Population":
-								cn = "population"
-							case "Continent":
-								cn = "continent"
-							case "tld":
-								cn = "tld"
-							case "CurrencyCode":
-								cn = "currency_code"
-							case "CurrencyName":
-								cn = "currency_name"
-							case "Phone":
-								cn = "phone"
-							case "Postal Code Format":
-								cn = "postal_format"
-							case "Postal Code Regex":
-								cn = "postal_regex"
-							case "Languages":
-								cn = "languages"
-							case "geonameid":
-								cn = "geonameId"
-							case "neighbours":
-								cn = "neighbours"
-							case "EquivalentFipsCode":
-								cn = "equivalent_fips_code"
-							default:
-								return errors.Errorf("Failed to map column %s", fn)
+						for _, row := range lines {
+							var args []interface{}
+							for _, v := range row {
+								args = append(args, v)
 							}
-							columns = append(columns, cn)
-						}
 
-						placeholders := []string{}
-						for i := 0; i < len(columns); i++ {
-							placeholders = append(placeholders, "?")
-						}
-
-						q := "insert into countryinfo (" + strings.Join(columns, ",") + ") values(" + strings.Join(placeholders, ",") + ")"
-						q = db.Rebind(q)
-						stmt, err = db.Prepare(q)
-						if err != nil {
-							return errors.WithMessagef(err, "Failed to prepare sql query '%s'", q)
+							_, err = stmt.Exec(args...)
+							if err != nil {
+								return errors.WithStack(err)
+							}
 						}
 					}
 
-					r := csv.NewReader(strings.NewReader(line))
-					r.Comma = '\t' // Use tab-delimited instead of comma <---- here!
-					r.FieldsPerRecord = -1
-
-					lines, err := r.ReadAll()
-					if err != nil {
+					if err := scanner.Err(); err != nil {
 						return errors.WithStack(err)
 					}
 
-					for _, row := range lines {
-						var args []interface{}
-						for _, v := range row {
-							args = append(args, v)
-						}
-
-						_, err = stmt.Exec(args...)
-						if err != nil {
-							return errors.WithStack(err)
-						}
-					}
-				}
-
-				if err := scanner.Err(); err != nil {
-					return errors.WithStack(err)
-				}
-
-				queries := []string{
-					// Countries...
-					`DROP TABLE IF EXISTS countries`,
-					`CREATE TABLE countries(
-						code           char(2) not null constraint countries_pkey primary key,
-						iso_alpha3           char(3),
-						name    character varying(50),
-						capital    character varying(50),
-						currency_code        char(3),
-						currency_name        CHAR(20),
-						phone                character varying(20),
-						postal_code_format        character varying(200),
-						postal_code_regex         character varying(200))`,
-					`insert into countries(code, iso_alpha3, name, capital, currency_code, currency_name, phone, postal_code_format, postal_code_regex)
+					queries := []string{
+						`insert into countries(code, iso_alpha3, name, capital, currency_code, currency_name, phone, postal_code_format, postal_code_regex)
 						select iso_alpha2, iso_alpha3, country, capital, currency_code, currency_name, phone, postal_format, postal_regex
 						from countryinfo`,
-					`DROP TABLE IF EXISTS countryinfo`,
-				}
+						`DROP TABLE IF EXISTS countryinfo`,
+					}
 
-				for _, q := range queries {
-					_, err := db.Exec(q)
-					if err != nil {
-						return errors.WithMessagef(err, "Failed to execute sql query '%s'", q)
+					for _, q := range queries {
+						_, err := db.Exec(q)
+						if err != nil {
+							return errors.WithMessagef(err, "Failed to execute sql query '%s'", q)
+						}
 					}
 				}
 
@@ -488,48 +507,52 @@ func migrationList(db *sqlx.DB, log *log.Logger) []*sqlxmigrate.Migration {
 					}
 				}
 
-				u := "http://download.geonames.org/export/dump/timeZones.txt"
-				resp, err := pester.Get(u)
-				if err != nil {
-					return errors.WithMessagef(err, "Failed to read timezones info from '%s'", u)
-				}
-				defer resp.Body.Close()
+				if isUnittest {
 
-				q := "insert into country_timezones (country_code,timezone_id) values(?, ?)"
-				q = db.Rebind(q)
-				stmt, err := db.Prepare(q)
-				if err != nil {
-					return errors.WithMessagef(err, "Failed to prepare sql query '%s'", q)
-				}
-
-				scanner := bufio.NewScanner(resp.Body)
-				for scanner.Scan() {
-					line := scanner.Text()
-
-					// Skip comments.
-					if strings.HasPrefix(line, "CountryCode") {
-						continue
-					}
-
-					r := csv.NewReader(strings.NewReader(line))
-					r.Comma = '\t' // Use tab-delimited instead of comma <---- here!
-					r.FieldsPerRecord = -1
-
-					lines, err := r.ReadAll()
+				} else {
+					u := "http://download.geonames.org/export/dump/timeZones.txt"
+					resp, err := pester.Get(u)
 					if err != nil {
-						return errors.WithStack(err)
+						return errors.WithMessagef(err, "Failed to read timezones info from '%s'", u)
+					}
+					defer resp.Body.Close()
+
+					q := "insert into country_timezones (country_code,timezone_id) values(?, ?)"
+					q = db.Rebind(q)
+					stmt, err := db.Prepare(q)
+					if err != nil {
+						return errors.WithMessagef(err, "Failed to prepare sql query '%s'", q)
 					}
 
-					for _, row := range lines {
-						_, err = stmt.Exec(row[0], row[1])
+					scanner := bufio.NewScanner(resp.Body)
+					for scanner.Scan() {
+						line := scanner.Text()
+
+						// Skip comments.
+						if strings.HasPrefix(line, "CountryCode") {
+							continue
+						}
+
+						r := csv.NewReader(strings.NewReader(line))
+						r.Comma = '\t' // Use tab-delimited instead of comma <---- here!
+						r.FieldsPerRecord = -1
+
+						lines, err := r.ReadAll()
 						if err != nil {
 							return errors.WithStack(err)
 						}
-					}
-				}
 
-				if err := scanner.Err(); err != nil {
-					return errors.WithStack(err)
+						for _, row := range lines {
+							_, err = stmt.Exec(row[0], row[1])
+							if err != nil {
+								return errors.WithStack(err)
+							}
+						}
+					}
+
+					if err := scanner.Err(); err != nil {
+						return errors.WithStack(err)
+					}
 				}
 
 				return nil
